# perf: cost-free conversion from paths to `&str`