--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -80,6 +80,8 @@ jobs:
           echo CARGO_REGISTRIES_CRATES_IO_PROTOCOL=git >> $GITHUB_ENV
       - name: Install cargo-hack
         uses: taiki-e/install-action@cargo-hack
+      - name: Install cargo-rm
+        run: cargo install cargo-edit --version 0.11.0
       - name: Use pinned Cargo.lock for Rust 1.34 and 1.44
         if: ${{ matrix.rust-version == 1.34 || matrix.rust-version == 1.44 }}
         run: cp Cargo.lock.rust134 Cargo.lock
@@ -87,7 +89,7 @@ jobs:
         run: cargo build
       # Remove criterion to support testing with earlier versions
       - name: Remove criterion
-        run: cargo remove --dev criterion
+        run: cargo-rm rm --dev criterion
       - name: Test
         run: cargo test
       - name: Build all targets with all features