--- a/Cargo.toml
+++ b/Cargo.toml
@@ -36,6 +36,7 @@ serde = { version = "1", features = ["derive"], optional = true }
 [dev-dependencies]
 bincode = "1"
 serde_bytes = "0.11.8"
+criterion = "0.4"
 
 [features]
 serde1 = ["serde"]