--- a/Cargo.toml
+++ b/Cargo.toml
@@ -21,10 +21,6 @@ authors = [
 edition = "2018"
 exclude = [".cargo/**/*", ".github/**/*"]
 
-[[bench]]
-name    = "bench"
-harness = false
-
 [package.metadata.docs.rs]
 all-features = true
 rustdoc-args = ["--cfg=doc_cfg"]
@@ -36,7 +32,6 @@ serde = { version = "1", features = ["derive"], optional = true }
 [dev-dependencies]
 bincode = "1"
 serde_bytes = "0.11.8"
-criterion = "0.4"
 
 [features]
 serde1 = ["serde"]