--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = [".", "camino-examples"]
+members = ["."]
 
 [package]
 name = "camino"