--- a/camino-examples/Cargo.toml
+++ b/camino-examples/Cargo.toml
@@ -8,7 +8,7 @@ publish = false
 [dependencies]
 anyhow = "1.0.38"
 camino = { path = "..", features = ["serde1"] }
-clap = { version = "4", features = ["derive"] }
+clap = { version = "3.0.7", features = ["derive"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = { version = "1.0.62" }