--- a/camino-examples/Cargo.toml
+++ b/camino-examples/Cargo.toml
@@ -1,3 +1,8 @@
+[workspace]
+# camino-examples pulls in crates that aren't supported by old versions of Rust, so make it its own
+# workspace.
+members = ["."]
+
 [package]
 name = "camino-examples"
 description = "Examples for camino"