… camino-examples/.gitignore