@h-a-n-a 2024-08-15
> would you be willing to update this PR with it

I would love to! Would you also like to integrate [codspeed](https://codspeed.io/) into the repo? If the answer is a yes, I can add a feature `"codspeed"` to enable the bench case running with codspeed on CI (This requires additional `CODSPEED_TOKEN` to get it running). Otherwise, I can just add the bench case.