@sunshowers 2024-08-15
This is now out in camino 1.1.8.